test: add tests for profile screens and button activation
