+++ app/src/main/java/com/github/se/cyrcle/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.github.se.cyrcle
 
-import CyrcleNavHost
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
