+++ app/src/main/java/com/github/se/cyrcle/ui/profile/ViewProfileScreen.kt
@@ -70,7 +70,7 @@ fun ViewProfileScreen(navigationActions: NavigationActions) {
       }
 
   Scaffold(
-      modifier = Modifier.testTag("ProfileScreen"),
+      modifier = Modifier.testTag("ViewProfileScreen"),
       bottomBar = {
         BottomNavigationBar(
             navigationActions = navigationActions,
