+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,8 @@
 <resources>
     <string name="app_name">Cyrcle</string>
     <string name="default_web_client_id">836297757379-3rvkggdu23fdvnof5nko2lablc993e0u.apps.googleusercontent.com</string>
+
+    <!-- Create Profile Screen -->
+    <string name="profile_screen_title">Create your profile</string>
+
 </resources>
